+++ app/src/main/java/com/example/leo3/data/model/RecordUiModel.kt
@@ -5,7 +5,7 @@ sealed class RecordUiModel {
     data class Header(
         val date: String,        // ex: 2025/12/08
         val weekDay: Int,        // 星期
-        val totalAmount: Int     // 當天所有明細的加總
+        val totalAmount: Long     // 當天所有明細的加總
     ) : RecordUiModel()
 
     data class Item(
